+++ src/lib/api/index.ts
@@ -1,6 +1,5 @@
 export * from "./types";
 export { createVoiceClone } from "./createVoiceClone";
-export { synthesizeDeepvoice } from "./synthesizeDeepvoice";
 export { createSession } from "./createSession";
 export { sendMessage } from "./sendMessage";
 export { createRealtimeCall } from "./createRealtimeCall";